Move `check_header` from `sp-consensus-subspace` into `sc-consensus-subspace`'s verifier with minimal changes
